requirements.txt
Django==1.6.5
Markdown==2.5
Twisted==12.2.0
boto==2.33.0
django-storages==1.1.8
djangorestframework==2.4.2
django-cors-headers==0.13
MySQL-python==1.2.5
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2